Rename navbar folder into sidebar
